+++ packages/polythene-react/scripts/standalone.js
@@ -1,7 +1,6 @@
 export * from "../index";
 export { subscribe, unsubscribe } from "polythene-core";
-import { addTypography } from "polythene-style";
-import { addLayoutStyles } from "polythene-core-css";
+import { addTypography, addLayoutStyles } from "polythene-css";
 import { addFastClick } from "polythene-fastclick";
 
 addTypography();