+++ packages/polythene-style/src/index.js
@@ -1,2 +1,1 @@
 export { vars } from "./variables";
-export { typographyStyles, addTypography } from "./typography";