+++ packages/tests-render-react/src/index.js
@@ -5,8 +5,7 @@ import {
   withRouter,
 } from "react-router-dom";
 import { addFastClick } from "polythene-fastclick";
-import { addTypography } from "polythene-style";
-import { addLayoutStyles } from "polythene-core-css";
+import { addTypography, addLayoutStyles } from "polythene-css";
 import "polythene-css";
 import { rules as css } from "./styles";
 import { renderer as h, List, Toolbar, ListTile, Snackbar, Notification } from "polythene-react";